Conflicts between the devtoolverse and other packages

This function lists all the conflicts between packages in the devtoolverse and other packages that you have loaded.

devtoolverse_conflicts()

Details

If dplyr is one of the select packages, then the following four conflicts are deliberately ignored: intersect, union, setequal, and setdiff from dplyr. These functions make the base equivalents generic, so shouldn't negatively affect any existing code.

Examples

devtoolverse_conflicts()

#> -- [1mConflicts[22m -------------------------------------------------------------------------------------- devtoolverse_conflicts() --
#> [31mx[39m [34mpkgdown[39m::[32mbuild_site()[39m          masks [34mdevtools[39m::build_site()
#> [31mx[39m [34mrhub[39m::[32mcheck()[39m                  masks [34mdevtools[39m::check()
#> [31mx[39m [34mlintr[39m::[32mlint()[39m                  masks [34mdevtools[39m::lint()
#> [31mx[39m [34mtestthat[39m::[32msetup()[39m              masks [34mdevtools[39m::setup()
#> [31mx[39m [34mtestthat[39m::[32mtest_file()[39m          masks [34mdevtools[39m::test_file()
#> [31mx[39m [34musethis[39m::[32muse_appveyor()[39m        masks [34mdevtools[39m::use_appveyor()
#> [31mx[39m [34musethis[39m::[32muse_build_ignore()[39m    masks [34mdevtools[39m::use_build_ignore()
#> [31mx[39m [34musethis[39m::[32muse_code_of_conduct()[39m masks [34mdevtools[39m::use_code_of_conduct()
#> [31mx[39m [34musethis[39m::[32muse_coverage()[39m        masks [34mdevtools[39m::use_coverage()
#> [31mx[39m [34musethis[39m::[32muse_cran_badge()[39m      masks [34mdevtools[39m::use_cran_badge()
#> [31mx[39m [34musethis[39m::[32muse_cran_comments()[39m   masks [34mdevtools[39m::use_cran_comments()
#> [31mx[39m [34musethis[39m::[32muse_data()[39m            masks [34mdevtools[39m::use_data()
#> [31mx[39m [34musethis[39m::[32muse_data_raw()[39m        masks [34mdevtools[39m::use_data_raw()
#> [31mx[39m [34musethis[39m::[32muse_dev_version()[39m     masks [34mdevtools[39m::use_dev_version()
#> [31mx[39m [34musethis[39m::[32muse_git()[39m             masks [34mdevtools[39m::use_git()
#> [31mx[39m [34musethis[39m::[32muse_git_hook()[39m        masks [34mdevtools[39m::use_git_hook()
#> [31mx[39m [34musethis[39m::[32muse_github()[39m          masks [34mdevtools[39m::use_github()
#> [31mx[39m [34musethis[39m::[32muse_github_links()[39m    masks [34mdevtools[39m::use_github_links()
#> [31mx[39m [34musethis[39m::[32muse_gpl3_license()[39m    masks [34mdevtools[39m::use_gpl3_license()
#> [31mx[39m [34musethis[39m::[32muse_mit_license()[39m     masks [34mdevtools[39m::use_mit_license()
#> [31mx[39m [34musethis[39m::[32muse_news_md()[39m         masks [34mdevtools[39m::use_news_md()
#> [31mx[39m [34musethis[39m::[32muse_package()[39m         masks [34mdevtools[39m::use_package()
#> [31mx[39m [34musethis[39m::[32muse_package_doc()[39m     masks [34mdevtools[39m::use_package_doc()
#> [31mx[39m [34musethis[39m::[32muse_rcpp()[39m            masks [34mdevtools[39m::use_rcpp()
#> [31mx[39m [34musethis[39m::[32muse_readme_md()[39m       masks [34mdevtools[39m::use_readme_md()
#> [31mx[39m [34musethis[39m::[32muse_readme_rmd()[39m      masks [34mdevtools[39m::use_readme_rmd()
#> [31mx[39m [34musethis[39m::[32muse_revdep()[39m          masks [34mdevtools[39m::use_revdep()
#> [31mx[39m [34musethis[39m::[32muse_rstudio()[39m         masks [34mdevtools[39m::use_rstudio()
#> [31mx[39m [34musethis[39m::[32muse_test()[39m            masks [34mdevtools[39m::use_test()
#> [31mx[39m [34musethis[39m::[32muse_testthat()[39m        masks [34mdevtools[39m::use_testthat()
#> [31mx[39m [34musethis[39m::[32muse_travis()[39m          masks [34mdevtools[39m::use_travis()
#> [31mx[39m [34musethis[39m::[32muse_vignette()[39m        masks [34mdevtools[39m::use_vignette()